+++ api/routes/user.php
@@ -55,7 +55,7 @@ $app->group('/user', function() use ($app) {
   });
 
   /**
-   * @api {post} /user/ Ajout d'un nouveau user.
+   * @api {post} /user/ Ajout d'un nouveau user. Sécurisé Admin.
    * @apiName PostUser
    * @apiGroup User
    *
@@ -74,17 +74,26 @@ $app->group('/user', function() use ($app) {
    *     }
    */
   $app->post('/',function ($request, $response)  use ($app) {
-    try {
-      Capsule::table('USER')->insert($request->getParsedBody());
-      $response = $response->withJson(array ("status"  => array("success" => "ok")), 200);
-    } catch(Illuminate\Database\QueryException $e) {
-      $response = $response->withJson(array ("status"  => array("error" => $e->getMessage())), 400);
+    if(isset($_SERVER['PHP_AUTH_USER']) && isset($_SERVER['HTTP_AUTHORIZATION'])){
+      $user = checkAuth($_SERVER['PHP_AUTH_USER'], $_SERVER['HTTP_AUTHORIZATION']);
+      if($user && $user->access == 1){
+        try {
+          Capsule::table('USER')->insert($request->getParsedBody());
+          $response = $response->withJson(array ("status"  => array("success" => "ok")), 200);
+        } catch(Illuminate\Database\QueryException $e) {
+          $response = $response->withJson(array ("status"  => array("error" => $e->getMessage())), 400);
+        }
+      }else{
+        $response = $response->withJson(array ("status"  => array("error" => "connexion")), 400);
+      }
+    }else{
+      $response = $response->withJson(array ("status"  => array("error" => "connexion")), 400);
     }
     return $response;
   });
 
   /**
-   * @api {put} /user/:login Modification d'un user.
+   * @api {put} /user/:login Modification d'un user. Sécurisé Admin.
    * @apiName PutUser
    * @apiGroup User
    *
@@ -103,17 +112,26 @@ $app->group('/user', function() use ($app) {
    *     }
    */
   $app->put('/{login}', function ($request, $response, $login) use ($app){
-    try {
-      Capsule::table('USER')->where('login',$login)->update($request->getParsedBody());
-      $response = $response->withJson(array ("status"  => array("success" => "ok")), 200);
-    } catch(Illuminate\Database\QueryException $e) {
-      $response = $response->withJson(array ("status"  => array("error" => $e->getMessage())), 400);
+    if(isset($_SERVER['PHP_AUTH_USER']) && isset($_SERVER['HTTP_AUTHORIZATION'])){
+      $user = checkAuth($_SERVER['PHP_AUTH_USER'], $_SERVER['HTTP_AUTHORIZATION']);
+      if($user && $user->access == 1){
+        try {
+          Capsule::table('USER')->where('login',$login)->update($request->getParsedBody());
+          $response = $response->withJson(array ("status"  => array("success" => "ok")), 200);
+        } catch(Illuminate\Database\QueryException $e) {
+          $response = $response->withJson(array ("status"  => array("error" => $e->getMessage())), 400);
+        }
+      }else{
+        $response = $response->withJson(array ("status"  => array("error" => "connexion")), 400);
+      }
+    }else{
+      $response = $response->withJson(array ("status"  => array("error" => "connexion")), 400);
     }
     return $response;
   });
 
   /**
-   * @api {delete} /user/:login Suppression d'un user.
+   * @api {delete} /user/:login Suppression d'un user. Sécurisé Mobile Admin.
    * @apiName DeleteUser
    * @apiGroup User
    *
@@ -132,11 +150,20 @@ $app->group('/user', function() use ($app) {
    *     }
    */
   $app->delete('/{login}',function ($request, $response, $login) {
-    try {
-      Capsule::table('COMMAND')->where('login',$login)->delete();
-      $response = $response->withJson(array ("status"  => array("success" => "ok")), 200);
-    } catch(Illuminate\Database\QueryException $e) {
-      $response = $response->withJson(array ("status"  => array("error" => $e->getMessage())), 400);
+    if(isset($_SERVER['PHP_AUTH_USER']) && isset($_SERVER['HTTP_AUTHORIZATION'])){
+      $user = checkAuth($_SERVER['PHP_AUTH_USER'], $_SERVER['HTTP_AUTHORIZATION']);
+      if($user && ($user->access == 1 || $user->access == 2)){
+        try {
+          Capsule::table('COMMAND')->where('login',$login)->delete();
+          $response = $response->withJson(array ("status"  => array("success" => "ok")), 200);
+        } catch(Illuminate\Database\QueryException $e) {
+          $response = $response->withJson(array ("status"  => array("error" => $e->getMessage())), 400);
+        }
+      }else{
+        $response = $response->withJson(array ("status"  => array("error" => "connexion")), 400);
+      }
+    }else{
+      $response = $response->withJson(array ("status"  => array("error" => "connexion")), 400);
     }
     return $response;
   });
